requirements.txt

fastapi==0.111.1
uvicorn[standard]==0.23.2
numpy<2
Pillow
tflite-runtime==2.14.0
python-multipart

